Fix CX gate overhead integration: update notebooks and GKSL simulator
- Fix cx_per_pair_gate type hint from int to float in qubit_noisy_simulator.py
- Update complete_comparison.ipynb Cell 15 to compute and use actual CX count
- Add cx_per_pair_gate support to QubitGKSLNoisyShotSimulator
- Update GKSL notebook Cell 38 to use cx_per_pair_gate with computed CX count
- Create iteration 8 verification script (30/30 checks pass)
- Create iteration 8 analysis document

+++ tutorials/qubit_noisy_simulator.py
@@ -348,7 +348,7 @@ class QubitMolecularDynamicsSimulatorNoisy:
                  initial_state_type: str = 'edge_triplet',
                  shots: int = 10000,
                  noise_params: Dict = None,
-                 cx_per_pair_gate: int = 1) -> Dict:
+                 cx_per_pair_gate: float = 1) -> Dict:
         """Run noisy simulation using density matrix with pair-level noise.
 
         Uses density matrix formalism with per-pair-gate 2-molecule depolarizing
